+++ example/android/app/src/main/kotlin/com/mohitkoley/flutter_app_functions_example/MyApplication.kt
@@ -1,5 +1,0 @@
-package com.mohitkoley.flutter_app_functions_example
-
-import com.mohitkoley.flutter_app_functions.FlutterAppFunctionsApplication
-
-class MyApplication : FlutterAppFunctionsApplication()